Refactor SessionAuth
