+++ packages/app/utils/supertokens/hooks/useAuthRedirect.ts
@@ -1,11 +1,11 @@
 import { useEffect } from 'react'
 import { useRouter } from 'solito/router'
-import { SessionWrapper } from '../SessionWrapper'
+import { sessionEventManager } from '../SessionEventManager'
 
 export const useAuthRedirect = () => {
   const router = useRouter()
   useEffect(() => {
-    const removeListener = SessionWrapper.getInstanceOrThrow().addEventListener((event) => {
+    const removeListener = sessionEventManager.addEventListener((event) => {
       if (event.action === 'SIGN_OUT' || event.action === 'UNAUTHORISED') {
         router.replace('/')
       }
